+++ esp32/snapclient/include/boot_mode_selector.h
@@ -3,3 +3,4 @@
 #include "snapclient_config.h"
 
 app_config::OperatingMode detectOperatingMode();
+void requestOperatingModeOnNextRestart(app_config::OperatingMode mode);