+++ esp32/snapclient/include/mode_switch_controller.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "snapclient_config.h"
+
+class ModeSwitchController {
+ public:
+  void begin(app_config::OperatingMode currentMode);
+  void update();
+
+ private:
+  bool isButtonActive() const;
+  void requestModeToggle();
+
+  app_config::OperatingMode currentMode_ = app_config::OperatingMode::Snapclient;
+  bool buttonWasReleasedAfterBoot_ = false;
+  bool pressHandled_ = false;
+  uint32_t pressedSinceMs_ = 0;
+};